Refactor: Moved chatbot to package, modal to module
Return value change
