+++ chatbot/__init__.py
@@ -1,0 +1,5 @@
+# A module that provides chatbot functionality to a Discord bot.
+# A chatbot is a query and response conversation between the bot and a user
+# designed to gather tabular information from them.
+
+from .chatbot import *
